Basil - Movement - RigidBody 2D
I created a test player, set up an input system that uses keyboard at the moment and made a PlayerMovement Script based on Brackey's in this video.https://youtu.be/whzomFgjT50.
I also tested it, and without the input manager it does not work, as it shows in the youtube video. I created a separate branch for this.