Fixes, and changes
Fixed Player health so that it was used, not a copy. Also fixed melt-timer so that it used a for loop instead of multiples of the same code. Changed floats related to health to Integers to account for new Health code. Also fixed flamethrower despawning issue, by adding a Destroy(gameobject, delay) unity function into the spawn code.