Follow Particle.
Added a trail article to Merlin, and adjusted it's duration from the default of 5 to 0.2. Also attached it to Merlin's wheel sprite in the prefab builder, as that is the "Source" of the particles. I also changed the colour from default white to blue, as that is a placeholder colour. This created a default Fade effect on Merlin, that I have not dealt with yet.