Particles added.
I added on Hit particles for Merlin and enemy entities.
Merlin Works fine on hit. Enemy Entities don't, as the Update function doesn't seem to work except to destroy the enemy it is attached to. I also added an OnSpawn artice system for Items. No troubleShooting necessary, as Items and weapons once picked up do not start using particles in the inventory.