Particle edits.
I changed the speed of the damage Particles from 60 to 20, then from 20 to 2. This fixed the screen being covered in particles. I also changed the smoke colors for Merlin to grey - dark grey and dark grey - black using two gradients. I also did this for the enemies. Item spawn particles did not have color changes, but did have the lifetime decreased to 0.5, and   the speed changed to 2.