Bean Weapon.
Created Bean Straw and Bean Projectiles.Created Bean Straw as Rapid Fire Projectile launcher, Except I upped the firing rate and downed the projectile speed.I created the Bean Projectile as a way of differenciating the weapon from other weapons at first, then pushed the idea forward.I used the concept Bean instead of Beam, when I mispelled Beam and thought it would be funny to make an OP Beam Weapon out of Beans. The images above are placeholder images, for I am not a Good Sprite Artist.I have not added any Safety features in, but I will recommend a timer on the firing system specifically for the beam weapon. It is at Minimum Damage of 1, and has a firing rate of 25. I have set the Projectile speed to 5, and the firing rate increases the effectiveness of the projectiles at that speed.
i decided to start with a Fully Automatic style weapon, rather than a pulse beam style weapon, to start as a placeholder, instead of creating new code at the moment.