Created Beam Weapon
Made a Beam Weapon that fires using a Boolean named _firing. Also created a Projectile Beam Behaviour class that is used by the beam weapon. The update was designed to create a range around the player where he is shooting a Beam. Currently it creates a range around the player only based on the transform.position of the spawn of the starting room.