Fixed logic error
The rooms were getting paused, but because Update() was getting called before the room was loaded to remove the previous room, the pause menu could be called because of the Resume function in Update. This was fixed by removing the Resume function from update, and moving it to after the player was moved.