Disabled Input logic from Game State.
I disabled Input logic, as 2 of these bugs were linked. 1 due to the input logic, the other I could not see because of the input logic. When I fixed the input "bug", I saw that the 2nd bug seemed to be resolved. As Aim-direction  is now set to the players last input device and location.