Fixed layering issue for particle.
To accomplish, I changed the Transform.position.z component to -0.2f on the prefab, and changed the spawn to -0.1f on the z. Due to prefabs not updating transforms, I am manually changing the z.position of existing enemies that are affected. This is not interfering with existing rooms at all.