Add settings menu
This is a large commit (sorry).

This adds a settings menu with two volume sliders. In order to
do this, changes to the audio manager were made that allow for
changing existing sounds based on the category (effect, music, etc.)
which then required modifying all usages of the Play() method,
hence the large list of files changed.